moved registration and custom auth forms into forms.py, ran makemigrations
